+++ tket/tests/tkettestsfiles.cmake
@@ -20,20 +20,20 @@ set(TEST_SOURCES
     # We should test simpler modules (e.g. Op, Circuit) before
     # the more complicated things that rely on them (e.g. Routing,
     # Transform) to help identify exactly where stuff breaks
-    # ${TKET_TESTS_DIR}/Utils/test_CosSinDecomposition.cpp
-    # ${TKET_TESTS_DIR}/Utils/test_HelperFunctions.cpp
-    # ${TKET_TESTS_DIR}/Utils/test_MatrixAnalysis.cpp
-    # ${TKET_TESTS_DIR}/Utils/test_RNG.cpp
-    # ${TKET_TESTS_DIR}/Graphs/test_GraphColouring.cpp
-    # ${TKET_TESTS_DIR}/Graphs/test_GraphFindComponents.cpp
-    # ${TKET_TESTS_DIR}/Graphs/test_GraphFindMaxClique.cpp
-    # ${TKET_TESTS_DIR}/Graphs/test_GraphUtils.cpp
-    # ${TKET_TESTS_DIR}/Graphs/test_DirectedGraph.cpp
-    # ${TKET_TESTS_DIR}/Graphs/test_ArticulationPoints.cpp
-    # ${TKET_TESTS_DIR}/Graphs/test_TreeSearch.cpp
-    # # NOTE: For testing TokenSwapping, it is easier to make use of
-    # # Architecture to set up test problems, rather than trying
-    # # to separate TokenSwapping-without-Architecture tests.
+    ${TKET_TESTS_DIR}/Utils/test_CosSinDecomposition.cpp
+    ${TKET_TESTS_DIR}/Utils/test_HelperFunctions.cpp
+    ${TKET_TESTS_DIR}/Utils/test_MatrixAnalysis.cpp
+    ${TKET_TESTS_DIR}/Utils/test_RNG.cpp
+    ${TKET_TESTS_DIR}/Graphs/test_GraphColouring.cpp
+    ${TKET_TESTS_DIR}/Graphs/test_GraphFindComponents.cpp
+    ${TKET_TESTS_DIR}/Graphs/test_GraphFindMaxClique.cpp
+    ${TKET_TESTS_DIR}/Graphs/test_GraphUtils.cpp
+    ${TKET_TESTS_DIR}/Graphs/test_DirectedGraph.cpp
+    ${TKET_TESTS_DIR}/Graphs/test_ArticulationPoints.cpp
+    ${TKET_TESTS_DIR}/Graphs/test_TreeSearch.cpp
+    # NOTE: For testing TokenSwapping, it is easier to make use of
+    # Architecture to set up test problems, rather than trying
+    # to separate TokenSwapping-without-Architecture tests.
     ${TKET_TESTS_DIR}/TokenSwapping/TableLookup/test_CanonicalRelabelling.cpp
     ${TKET_TESTS_DIR}/TokenSwapping/TableLookup/test_ExactMappingLookup.cpp
     ${TKET_TESTS_DIR}/TokenSwapping/TableLookup/test_FilteredSwapSequences.cpp
@@ -63,7 +63,6 @@ set(TEST_SOURCES
     ${TKET_TESTS_DIR}/Circuit/test_Circ.cpp
     ${TKET_TESTS_DIR}/Circuit/test_Symbolic.cpp
     ${TKET_TESTS_DIR}/Circuit/test_ThreeQubitConversion.cpp
-    ${TKET_TESTS_DIR}/test_Program.cpp
     ${TKET_TESTS_DIR}/test_CliffTableau.cpp
     ${TKET_TESTS_DIR}/test_UnitaryTableau.cpp
     ${TKET_TESTS_DIR}/test_PhasePolynomials.cpp